Install Python in GHA deployment runner
Without this step, the pyproject.toml file is not correctly parsed, see
e.g. this action which fails:

https://github.com/Urban-Analytics-Technology-Platform/web-app-template/actions/runs/9894661493/job/27332702540#step:7:108

Notice that the package name is UNKNOWN and the wheel file is
UNKNOWN-0.0.0-py3-none-any.whl. This causes the web template to break
because it expects my-python-module.<stuff>.whl.

I think this is essentially an extension of
https://github.com/pypa/setuptools/issues/3269. Installing a proper
version of Python (which presumably includes pip and setuptools) fixes
this issue.